restructure the prj and add methods to read/write books from/in json files
