+++ classes/nameable.rb
